backend/requirements.txt
Flask
pycryptodome
python-docx
flask-cors
